Fix theme color toggler
Color theme was in the header section of the page was visually broken.
When a user clicks on the toggler, the website's theme color is supposed
to change with the toggle buttin animating between left and right.
And here was the case where the theme color changed when the user clicks in
the button but the toggle animation exhibited unexpected behaviour